/finalizando

+++ Calendly/src/Api/Service/Service.cs
@@ -8,54 +8,53 @@ public class Serv : IServ
 {
     private readonly IRepo _repository = new Repo();
     
-    public List<EventDto> ListEvents ()
+    public List<EventGetDto> ListEvents ()
     {
-        List<EventDto> events = new ();
+        List<EventGetDto> events = new ();
         
         _repository.ListEvents().ForEach(
-            eventDao => events.Add(new EventDto(eventDao))
+            eventDao => events.Add(new EventGetDto(eventDao))
         );
         
         return events;
     }
 
-    public EventDto FindEventById(string uid)
+    public EventGetDto? FindEventById(string uid)
     {
-        EventDto eventDto = new EventDto(_repository.FindEventById(uid));
-
-        return eventDto;
+        EventDao? eventDaoFound = _repository.FindEventById(uid);
+        return eventDaoFound is not null ? new EventGetDto(eventDaoFound) : null;
     }
 
-    public string AddEvent(string eventName, int eventDuration, string eventLocation, string eventDescription)
+    public string AddEvent(EventPostDto newEventDto)
     {
-        EventDto newEvent = new EventDto(eventName, eventDuration, eventLocation, eventDescription);
+        if (newEventDto.EventDuration <= 0) return "Deve-se inserir a duração do Evento!";
+
+        IEventDto newEvent = new EventPostDto(
+            eventName: newEventDto.EventName,
+            eventDuration: newEventDto.EventDuration,
+            eventLocation: newEventDto.EventLocation
+        );
+
         bool result = _repository.AddEvent(newEvent);
-        
+
         return result?"Criado com Sucesso":"Falha na Criação";
     }
 
-    public string UpdateEvent(string eventUId, EventDtoPut newEventDto)
+    public string UpdateEvent(string eventUId, EventPutDto newEventDto)
     {
-        EventDto eventToUpdateDto = FindEventById(eventUId);
+        IEventDto? eventToUpdateDto = FindEventById(eventUId);
 
-        // ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
-        if (eventToUpdateDto is null)
-        {
-            return "Evento não encontrado";
-        }
+        if (eventToUpdateDto is null) return "Evento não encontrado";
+        if (newEventDto.EventDuration <= 0) return "Deve-se inserir a duração do Evento!";
 
         foreach (var property in newEventDto.GetType().GetProperties())
         {
             var propertyValue = property.GetValue(newEventDto);
-
-            if (property.Name == "EventLastUpdateTime")
-            {
-                eventToUpdateDto.GetType().GetProperty(property.Name)!.SetValue(eventToUpdateDto, DateTime.Now);
-            } else if (propertyValue != null)
-            {
-                eventToUpdateDto.GetType().GetProperty(property.Name)!.SetValue(eventToUpdateDto, propertyValue);
-            }
+            if (propertyValue != null) eventToUpdateDto.GetType().GetProperty(property.Name)!
+                .SetValue(eventToUpdateDto, propertyValue);
         }
+
+        eventToUpdateDto.EventLastUpdateTime = DateTime.UtcNow;
 
         bool result = _repository.UpdateEvent(eventUId, eventToUpdateDto);
 
@@ -64,9 +63,12 @@ public class Serv : IServ
 
     public string DeleteEvent(string eventUId)
     {
+
+        IEventDto? eventToDeleteDto = FindEventById(eventUId);
+
+        if (eventToDeleteDto is null) return "Evento não encontrado";
+
         bool result = _repository.DeleteEvent(eventUId);
-        
         return result?"Deletado com Sucesso":"Falha na Deleção";
     }
-
 }
